+++ api/tests/unit/integrations/slack/conftest.py
@@ -1,0 +1,13 @@
+import pytest
+
+
+@pytest.fixture
+def mocked_slack_internal_client(mocker):
+    mocked_client = mocker.MagicMock()
+
+    mocker.patch(
+        "integrations.slack.slack.SlackWrapper._client",
+        new_callable=mocker.PropertyMock,
+        return_value=mocked_client,
+    )
+    return mocked_client
