+++ api/tests/unit/integrations/slack/test_unit_slack.py
@@ -1,18 +1,4 @@
-import pytest
-
 from integrations.slack.slack import SlackWrapper
-
-
-@pytest.fixture
-def mocked_slack_internal_client(mocker):
-    mocked_client = mocker.MagicMock()
-
-    mocker.patch(
-        "integrations.slack.slack.SlackWrapper._client",
-        new_callable=mocker.PropertyMock,
-        return_value=mocked_client,
-    )
-    return mocked_client
 
 
 def test_get_channels_data_response_structure(mocker, mocked_slack_internal_client):
